requirements.txt
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
future==0.18.3
greenlet==2.0.1
idna==3.4
Jinja2==3.1.2
loguru==0.6.0
lxml==4.9.2
MarkupSafe==2.1.2
netmiko==4.1.2
ntc-templates==3.3.0
paramiko==3.1.0
Pillow==9.4.0
playwright==1.32.1
pycparser==2.21
pyee==9.0.4
PyNaCl==1.5.0
pyserial==3.5
python-dotenv==1.0.0
python-pptx==0.6.21
PyYAML==6.0
requests==2.28.2
scp==0.14.5
six==1.16.0
tenacity==8.2.2
textfsm==1.1.2
typing_extensions==4.5.0
Unidecode==1.3.6
urllib3==1.26.15
XlsxWriter==3.0.9
